requirements.txt
streamlit==1.26.0
matplotlib-3.9.0